Rewrite test/fs_utils in mocha.
